Ai_model/requirements.txt
Flask
Flask-CORS
torch
torchvision
timm
numpy
Pillow
requests
gunicorn